requirements.txt
databento>=0.30.0
pandas>=2.0.0
numpy>=1.24.0
py_vollib>=1.0.1
yfinance>=0.2.0
matplotlib>=3.7.0
plotly>=5.15.0
pyyaml>=6.0
pytest>=7.0.0
pytest-cov>=7.0.0
alpaca-py>=0.21.0
ib_insync>=0.9.86
streamlit>=1.30.0
